util: Put function definitions in their own TU
